Refactor: Refactor Nav component
* Integrate unnecessary components

* Change location of Nav directory

* Add QuotesGroupType
