--- src/components/Main/Nav/NavContentWrapper/NoContents.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { StyledNoContents } from './StyledNoContents';
-
-const NoContents = () => {
-    return (
-        <StyledNoContents>
-            <span>No Contents</span>
-        </StyledNoContents>
-    );
-};
-
-export default NoContents;
