--- src/components/Main/Nav/NavContentWrapper/StyledNoContents.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import { navBoxMixin } from '~/styles/mixins/navBoxMixin';
-
-export const StyledNoContents = styled.li`
-    ${navBoxMixin}
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-`;
