+++ src/types/type.ts
@@ -3,3 +3,5 @@ export type ProviderNames = "Google" | "Facebook" | "Github";
 export type ChevronEventType = "prev" | "next";
 
 export type CardPositionType = "front" | "back";
+
+export type QuotesGroupType = 'history' | 'favorite';
